refactor menu bar code
