+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/Gui.java
@@ -26,7 +26,7 @@ import org.quiltmc.enigma.gui.docker.ObfuscatedClassesDocker;
 import org.quiltmc.enigma.gui.docker.StructureDocker;
 import org.quiltmc.enigma.gui.element.EditorTabbedPane;
 import org.quiltmc.enigma.gui.element.MainWindow;
-import org.quiltmc.enigma.gui.element.MenuBar;
+import org.quiltmc.enigma.gui.element.menu_bar.MenuBar;
 import org.quiltmc.enigma.gui.panel.EditorPanel;
 import org.quiltmc.enigma.gui.panel.IdentifierPanel;
 import org.quiltmc.enigma.gui.renderer.MessageListCellRenderer;
@@ -282,7 +282,7 @@ public class Gui {
 
 	public void addCrash(Throwable t) {
 		this.crashHistory.add(t);
-		this.menuBar.prepareCrashHistoryMenu();
+		this.menuBar.getFileMenu().prepareCrashHistoryMenu();
 	}
 
 	public DockerManager getDockerManager() {
