+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/element/menu_bar/AbstractEnigmaMenu.java
@@ -1,0 +1,6 @@
+package org.quiltmc.enigma.gui.element.menu_bar;
+
+import javax.swing.JMenu;
+
+public class AbstractEnigmaMenu extends JMenu implements EnigmaMenu {
+}
