+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/element/menu_bar/EnigmaMenu.java
@@ -1,0 +1,9 @@
+package org.quiltmc.enigma.gui.element.menu_bar;
+
+public interface EnigmaMenu {
+	default void setKeyBinds() {}
+
+	default void updateState() {}
+
+	default void retranslate() {}
+}
