+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/element/menu_bar/HelpMenu.java
@@ -1,0 +1,36 @@
+package org.quiltmc.enigma.gui.element.menu_bar;
+
+import org.quiltmc.enigma.gui.Gui;
+import org.quiltmc.enigma.gui.dialog.AboutDialog;
+import org.quiltmc.enigma.gui.util.GuiUtil;
+import org.quiltmc.enigma.util.I18n;
+
+import javax.swing.JMenuItem;
+
+public class HelpMenu extends AbstractEnigmaMenu {
+	private final Gui gui;
+
+	private final JMenuItem aboutItem = new JMenuItem();
+	private final JMenuItem githubItem = new JMenuItem();
+
+	public HelpMenu(Gui gui) {
+		this.gui = gui;
+
+		this.add(this.aboutItem);
+		this.add(this.githubItem);
+
+		this.aboutItem.addActionListener(e -> AboutDialog.show(this.gui.getFrame()));
+		this.githubItem.addActionListener(e -> this.onGithubClicked());
+	}
+
+	@Override
+	public void retranslate() {
+		this.setText(I18n.translate("menu.help"));
+		this.aboutItem.setText(I18n.translate("menu.help.about"));
+		this.githubItem.setText(I18n.translate("menu.help.github"));
+	}
+
+	private void onGithubClicked() {
+		GuiUtil.openUrl("https://github.com/QuiltMC/Enigma");
+	}
+}
